topla-backend/clip-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sentence-transformers==3.0.0
Pillow==10.4.0
python-multipart==0.0.9
numpy==1.26.4
torch==2.3.0+cpu -f https://download.pytorch.org/whl/torch_stable.html
